Add a template makefile for game projects
The template makefile is meant to be a starting point for game projects
to hook with/use the Jimara engine. On most linux systems it should not
require major changes, with notable exception being the necessity to
specify `JIMARA_REPO` variable that should be the path pointing to
Jimara reposity on the host and `GAME_SOURCE_DIR` if source files are
not under `./src`.